feat(api): add structured error handling and response utilities
