+++ packages/utils/api/client-utils.ts
@@ -1,9 +1,8 @@
 "use client";
-import { ApiError } from "@ayasofyazilim/core-saas/AccountService";
-import { notFound, permanentRedirect, RedirectType } from "next/navigation";
 import { toast } from "@/components/ui/sonner";
+import { ApiError } from "@ayasofyazilim/core-saas/AccountService";
 import type { AppRouterInstance } from "next/dist/shared/lib/app-router-context.shared-runtime";
-import { ApiErrorServerResponse, ServerResponse } from "./types";
+import { ServerResponse } from "./types";
 
 export const handlePutResponse = (
   response: { type: "success" | "api-error"; message: string },
@@ -74,66 +73,4 @@ export function isApiError(error: unknown): error is ApiError {
     return true;
   }
   return error instanceof ApiError;
-}
-
-export function structuredError(error: unknown): ApiErrorServerResponse {
-  if (isApiError(error)) {
-    const body = error.body as
-      | {
-          error: { message?: string; details?: string };
-        }
-      | undefined;
-    const errorDetails = body?.error || {};
-    return {
-      type: "api-error",
-      data: errorDetails.message || error.statusText || "Something went wrong",
-      message:
-        errorDetails.details ||
-        errorDetails.message ||
-        error.statusText ||
-        "Something went wrong",
-    };
-  }
-  return {
-    type: "api-error",
-    message: "[Unknown] Something went wrong",
-    data: "[Unknown] Something went wrong",
-  };
-}
-
-export function structuredResponse<T>(data: T): ServerResponse<T> {
-  return {
-    type: "success",
-    data,
-    message: "",
-  };
-}
-
-export function isErrorOnRequest<T>(
-  response: ServerResponse<T>,
-  lang: string,
-  redirectToNotFound = true,
-): response is {
-  type: "api-error";
-  message: string;
-  data: string;
-} {
-  if (response.type === "success") return false;
-
-  if (response.data === "Forbidden") {
-    return permanentRedirect(`/${lang}/unauthorized`, RedirectType.replace);
-  }
-
-  if (redirectToNotFound) {
-    return notFound();
-  }
-  return true;
-}
-
-export function structuredSuccessResponse<T>(data: T) {
-  return {
-    type: "success" as const,
-    data,
-    message: "",
-  };
 }
