+++ packages/utils/auth/auth.ts
@@ -9,7 +9,6 @@ import {
 } from "./auth-actions";
 import NextAuth, { AuthError } from "next-auth";
 import { MyUser } from "./auth-types";
-import { GetApiAbpApplicationConfigurationResponse } from "@ayasofyazilim/core-saas/AccountService";
 
 export const { handlers, auth, signIn, signOut } = NextAuth({
   providers: [
