+++ packages/utils/policies/utils.ts
@@ -1,7 +1,7 @@
 "use server";
 import { permanentRedirect, RedirectType } from "next/navigation";
 import { Policy } from "./types";
-import { getGrantedPoliciesApi } from "../api/action";
+import { getGrantedPoliciesApi } from "../api/api-action";
 export async function isUnauthorized({
   requiredPolicies,
   lang,
